feat: Phase 2 + Phase 3 complete -- 155 Indian stations, full pipeline, Core Feature UI
Phase 2:
- Expanded city_registry to 45 Indian cities, city_stations to 155 monitoring stations
- Built WAQIStationConnectorSubject polling all 155 stations every 180s
- OpenMeteoConnectorSubject for weather + forecast
- Pathway pipeline: join, classify_aqi UDF, postgres write, Redis pub/sub
- FastAPI Socket.IO server with Redis bridge

Phase 3 (Core Feature UI):
- station_aqi_history schema expanded with 13 new columns (pollutants + weather)
- 7 new API routes: cities, summary, stations, history, weather, pollutant, health-recommendation
- Frontend: Zustand store with Socket.IO live updates
- Components: AQIGauge, HealthInsightCard, WeatherCard, ForecastChart, StationDetailPanel, PollutantTabBar, CitySelector, Navbar
- Pages: Dashboard, Heatmap (Leaflet drill-down), PollutantPage
- React Router v7 layout with full routing
- Fixed: react-leaflet@4 for React 18 compat, Leaflet CSS global import, react-is dep
- Poll interval: 180s for prototyping

+++ backend/pathway_pipeline/connectors/waqi_station_connector.py
@@ -27,7 +27,7 @@ class WAQIStationConnectorSubject(pw.io.python.ConnectorSubject):
                             "city_id": station["city_id"],
                             "station_id": station["id"],
                             "city": station["station_name"],
-                            "aqi": int(d.get("aqi", 0) if str(d.get("aqi", 0)).isdigit() else 0),
+                            "aqi": float(d.get("aqi", 0) if str(d.get("aqi", 0)).isdigit() else 0),
                             "pm25": float(iaqi.get("pm25", {}).get("v", 0.0)),
                             "pm10": float(iaqi.get("pm10", {}).get("v", 0.0)),
                             "no2": float(iaqi.get("no2", {}).get("v", 0.0)),
